+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   OpenAI.
 - Now using `integer` as the type for integers in JSON schema.
 - Allow stripping out JSON schema extensions by option.
+- More strictly following JSON schema for nullable.
 
 ## 1.9.0